Move directives to Helper, rewrite directive string detection
